% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{annotator}
\alias{annotator}
\title{ggplot inlays}
\description{
Effortlessly Add ggplot Inlays with Shiny and patchwork::inset_element() for Simple Placement
}
\author{
Mihai Valcu \email{Mihai.Valcu@bi.mpg.de}
}
