% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gginlay.R
\name{gginlay}
\alias{gginlay}
\title{gginlay
gginlay}
\usage{
gginlay(g, p)
}
\description{
gginlay
gginlay
}
\examples{
g = ggplot(data=iris,aes(y=Sepal.Width, x= Species)) +
    geom_boxplot()
versicolor = system.file("iris", "versicolor.jpg", package = "gginlay") |> jpeg::readJPEG(native = TRUE)
gginlay(g, versicolor)
}
